Split jobs page helpers and rail
