+++ frontend/app/(core)/jobs/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { ChevronDown, Plus } from 'lucide-react';
+import { ChevronDown } from 'lucide-react';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { HeaderBar } from '@/components/HeaderBar';
 import { AppSidebar } from '@/components/AppSidebar';
@@ -16,162 +16,16 @@ import type { MediaLightboxEntry } from '@/components/MediaLightbox';
 import { normalizeGroupSummaries, normalizeGroupSummary } from '@/lib/normalize-group-summary';
 import dynamic from 'next/dynamic';
 import { useI18n } from '@/lib/i18n/I18nProvider';
-import { isPlaceholderMediaUrl } from '@/lib/media';
-import { AudioWaveformThumb } from '@/components/ui/AudioWaveformThumb';
-import {
-  resolveJobsRailAudio,
-  resolveJobsRailPlaceholderThumb,
-  resolveJobsRailThumb,
-  resolveJobsRailVideo,
-} from '@/lib/jobs-rail-thumb';
 import { Button } from '@/components/ui/Button';
-import type { Job } from '@/types/jobs';
-import type { JobSurface } from '@/types/billing';
-import { deriveJobSurface } from '@/lib/job-surface';
-
-const DEFAULT_JOBS_COPY = {
-  title: 'History',
-  sections: {
-    video: 'Video history',
-    audio: 'Audio history',
-    image: 'Image history',
-    character: 'Character history',
-    angle: 'Angle history',
-    upscale: 'Upscale history',
-    videoEmpty: 'No video renders yet.',
-    audioEmpty: 'No audio renders yet.',
-    imageEmpty: 'No image renders yet.',
-    characterEmpty: 'No character renders yet.',
-    angleEmpty: 'No angle renders yet.',
-    upscaleEmpty: 'No upscale renders yet.',
-  },
-  teams: {
-    title: 'Teams',
-    srLive: 'Live',
-    srSoon: 'Coming soon',
-    live: 'Shared wallets, approvals, and budgets are enabled in your workspace.',
-    beta: 'Coming soon — shared wallets, approvals, and budgets. Join the beta at {email}.',
-    email: 'support@maxvideoai.com',
-  },
-  curated: 'Starter renders curated by the MaxVideo team appear here until you generate your own clips.',
-  error: 'Failed to load history.',
-  retry: 'Retry',
-  empty: 'No renders yet. Start a generation to populate your history.',
-  loadMore: 'Load more',
-  loading: 'Loading…',
-  actions: {
-    addToLibrary: 'Add to Library',
-    saving: 'Saving…',
-    recreate: 'Generate same settings',
-    openDetails: 'View details',
-    actions: 'Actions',
-    expandSection: 'Show',
-    collapseSection: 'Hide',
-  },
-} as const;
-
-type JobsCopy = typeof DEFAULT_JOBS_COPY;
-type LibrarySaveKind = 'image' | 'video' | 'audio';
-type LibrarySavePayload = {
-  url: string;
-  kind: LibrarySaveKind;
-  jobId?: string | null;
-  label?: string | null;
-  thumbUrl?: string | null;
-  previewUrl?: string | null;
-};
-
-function resolveClientJobSurface(job: Job): JobSurface {
-  return deriveJobSurface({
-    surface: job.surface,
-    settingsSnapshot: job.settingsSnapshot,
-    jobId: job.jobId,
-    engineId: job.engineId ?? null,
-    videoUrl: job.videoUrl ?? null,
-    renderIds: job.renderIds,
-  });
-}
-
-function resolveWorkspaceJobHref(jobId: string, surface: JobSurface, forceImageGroup = false): string {
-  if (surface === 'audio') {
-    return `/app/audio?job=${encodeURIComponent(jobId)}`;
-  }
-  if (forceImageGroup || surface === 'image') {
-    return `/app/image?job=${encodeURIComponent(jobId)}`;
-  }
-  if (surface === 'upscale') {
-    return `/app/tools/upscale?job=${encodeURIComponent(jobId)}`;
-  }
-  return `/app?job=${encodeURIComponent(jobId)}`;
-}
-
-function firstHttpUrl(values: Array<string | null | undefined>): string | null {
-  return values.find((value): value is string => typeof value === 'string' && /^https?:\/\//i.test(value)) ?? null;
-}
-
-function resolveGroupLibrarySavePayload(group: GroupSummary): LibrarySavePayload | null {
-  const job = group.hero.job;
-  const members = group.members;
-  const firstPreview = group.previews[0];
-  const jobRenderUrl = firstHttpUrl(job?.renderIds ?? []);
-  const videoUrl = firstHttpUrl([group.hero.videoUrl, firstPreview?.videoUrl, ...members.map((member) => member.videoUrl)]);
-  const audioUrl = firstHttpUrl([group.hero.audioUrl, ...members.map((member) => member.audioUrl)]);
-  const thumbUrl = firstHttpUrl([group.hero.thumbUrl, firstPreview?.thumbUrl, job?.thumbUrl]);
-  const previewUrl = firstHttpUrl([group.hero.previewVideoUrl, firstPreview?.previewVideoUrl, job?.previewVideoUrl]);
-  const imageUrl = firstHttpUrl([jobRenderUrl, ...members.map((member) => member.thumbUrl), thumbUrl]);
-  const jobId = job?.jobId ?? group.hero.jobId ?? group.id;
-  const label = job?.prompt ?? group.hero.prompt ?? undefined;
-
-  if (videoUrl) {
-    return { url: videoUrl, kind: 'video', jobId, label, thumbUrl, previewUrl };
-  }
-  if (audioUrl) {
-    return { url: audioUrl, kind: 'audio', jobId, label, thumbUrl, previewUrl: null };
-  }
-  if (imageUrl) {
-    return { url: imageUrl, kind: 'image', jobId, label, thumbUrl, previewUrl: null };
-  }
-  return null;
-}
-
-function resolveEntryLibrarySavePayload(entry: MediaLightboxEntry): LibrarySavePayload | null {
-  const videoUrl = firstHttpUrl([entry.videoUrl]);
-  const audioUrl = firstHttpUrl([entry.audioUrl]);
-  const imageUrl = firstHttpUrl([entry.imageUrl, entry.thumbUrl]);
-  const label = entry.prompt ?? entry.label ?? undefined;
-
-  if (videoUrl) {
-    return {
-      url: videoUrl,
-      kind: 'video',
-      jobId: entry.jobId,
-      label,
-      thumbUrl: entry.thumbUrl ?? null,
-      previewUrl: entry.previewUrl ?? null,
-    };
-  }
-  if (audioUrl) {
-    return {
-      url: audioUrl,
-      kind: 'audio',
-      jobId: entry.jobId,
-      label,
-      thumbUrl: entry.thumbUrl ?? null,
-      previewUrl: null,
-    };
-  }
-  if (imageUrl) {
-    return {
-      url: imageUrl,
-      kind: 'image',
-      jobId: entry.jobId,
-      label,
-      thumbUrl: entry.thumbUrl ?? imageUrl,
-      previewUrl: null,
-    };
-  }
-  return null;
-}
+import { CollapsedGroupRail, CollapsedGroupRailSkeleton } from './_components/jobs-collapsed-rail';
+import { renderSkeletonCards } from './_components/jobs-skeleton-cards';
+import { DEFAULT_JOBS_COPY, type JobsCopy } from './_lib/jobs-copy';
+import {
+  resolveClientJobSurface,
+  resolveEntryLibrarySavePayload,
+  resolveGroupLibrarySavePayload,
+  resolveWorkspaceJobHref,
+} from './_lib/jobs-page-helpers';
 
 export default function JobsPage() {
   const { t } = useI18n();
@@ -843,162 +697,6 @@ export default function JobsPage() {
   );
 }
 
-function renderSkeletonCards(count: number, prefix: string) {
-  return Array.from({ length: count }).map((_, index) => (
-    <div key={`${prefix}-skeleton-${index}`} className="rounded-card border border-border bg-surface-glass-60 p-0" aria-hidden>
-      <div className="relative overflow-hidden rounded-card">
-        <div className="relative" style={{ aspectRatio: '16 / 9' }}>
-          <div className="skeleton absolute inset-0" />
-        </div>
-      </div>
-      <div className="border-t border-border bg-surface-glass-70 px-3 py-2">
-        <div className="h-3 w-24 rounded-full bg-skeleton" />
-      </div>
-    </div>
-  ));
-}
-const COLLAPSED_RAIL_ITEM_WIDTH = 220;
-
-function RailThumb({ src, videoSrc, fallbackSrc }: { src: string; videoSrc?: string | null; fallbackSrc: string }) {
-  const [resolvedSrc, setResolvedSrc] = useState(src);
-  const [imageFailed, setImageFailed] = useState(false);
-  const [videoFailed, setVideoFailed] = useState(false);
-
-  useEffect(() => {
-    setResolvedSrc(src);
-    setImageFailed(false);
-    setVideoFailed(false);
-  }, [src, videoSrc]);
-
-  const showVideo = Boolean(videoSrc && (isPlaceholderMediaUrl(resolvedSrc) || imageFailed) && !videoFailed);
-
-  if (showVideo) {
-    return (
-      <video
-        src={videoSrc ?? undefined}
-        poster={fallbackSrc}
-        className="absolute inset-0 h-full w-full object-cover"
-        muted
-        playsInline
-        loop
-        autoPlay
-        preload="metadata"
-        onError={() => {
-          setVideoFailed(true);
-          setResolvedSrc(fallbackSrc);
-        }}
-      />
-    );
-  }
-
-  return (
-    // eslint-disable-next-line @next/next/no-img-element
-    <img
-      src={resolvedSrc}
-      alt=""
-      className="absolute inset-0 h-full w-full object-cover"
-      loading="lazy"
-      onError={() => {
-        setImageFailed(true);
-        if (!videoSrc && resolvedSrc !== fallbackSrc) {
-          setResolvedSrc(fallbackSrc);
-        }
-      }}
-    />
-  );
-}
-
-function CollapsedGroupRailSkeleton() {
-  return (
-    <div className="flex gap-4 overflow-x-auto pb-2">
-      {Array.from({ length: 8 }).map((_, index) => (
-        <div
-          key={`jobs-rail-skeleton-${index}`}
-          className="relative shrink-0 overflow-hidden rounded-card border border-border bg-surface-glass-60"
-          style={{ width: COLLAPSED_RAIL_ITEM_WIDTH }}
-          aria-hidden
-        >
-          <div className="relative" style={{ aspectRatio: '16 / 9' }}>
-            <div className="skeleton absolute inset-0" />
-          </div>
-        </div>
-      ))}
-    </div>
-  );
-}
-
-function CollapsedGroupRail({
-  groups,
-  onOpen,
-  onSaveToLibrary,
-  savingIds,
-  addLabel,
-  savingLabel,
-}: {
-  groups: GroupSummary[];
-  onOpen: (group: GroupSummary) => void;
-  onSaveToLibrary: (group: GroupSummary) => void;
-  savingIds: Set<string>;
-  addLabel: string;
-  savingLabel: string;
-}) {
-  const items = groups.slice(0, 12);
-  return (
-    <div className="flex gap-4 overflow-x-auto pb-2">
-      {items.map((group) => {
-        const thumb = resolveJobsRailThumb(group);
-        const fallbackThumb = resolveJobsRailPlaceholderThumb(group.hero.aspectRatio ?? group.previews[0]?.aspectRatio ?? null);
-        const video = resolveJobsRailVideo(group);
-        const audio = resolveJobsRailAudio(group);
-        const isSaving = savingIds.has(group.id);
-        return (
-          <div
-            key={group.id}
-            className="group relative block h-auto min-h-0 shrink-0 overflow-hidden rounded-card border border-border bg-surface p-0 shadow-card transition hover:border-border-hover focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 focus-visible:ring-offset-bg"
-            style={{ width: COLLAPSED_RAIL_ITEM_WIDTH }}
-          >
-            <button
-              type="button"
-              onClick={() => onOpen(group)}
-              className="block w-full"
-              aria-label="Open render"
-            >
-              <div className="relative w-full overflow-hidden" style={{ aspectRatio: '16 / 9' }}>
-                {video ? (
-                  <RailThumb src={thumb} videoSrc={video} fallbackSrc={fallbackThumb} />
-                ) : audio ? (
-                  <AudioWaveformThumb
-                    seed={audio}
-                    thumbSrc={thumb !== fallbackThumb ? thumb : null}
-                    label={null}
-                    active={false}
-                  />
-                ) : (
-                  <RailThumb src={thumb} videoSrc={video} fallbackSrc={fallbackThumb} />
-                )}
-                {group.count > 1 ? (
-                  <div className="absolute bottom-2 right-2 rounded-full bg-surface-on-media-dark-55 px-2 py-0.5 text-xs font-semibold text-on-inverse">
-                    ×{group.count}
-                  </div>
-                ) : null}
-              </div>
-            </button>
-            <button
-              type="button"
-              onClick={() => onSaveToLibrary(group)}
-              disabled={isSaving}
-              title={isSaving ? savingLabel : addLabel}
-              aria-label={isSaving ? savingLabel : addLabel}
-              className="absolute right-2 top-2 flex h-8 w-8 items-center justify-center rounded-full border border-white/70 bg-white/90 text-black/80 shadow-md backdrop-blur transition hover:bg-white focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring disabled:cursor-wait disabled:opacity-70"
-            >
-              <Plus className="h-4 w-4" aria-hidden="true" />
-            </button>
-          </div>
-        );
-      })}
-    </div>
-  );
-}
 const GroupViewerModal = dynamic(
   () => import('@/components/groups/GroupViewerModal').then((mod) => mod.GroupViewerModal),
   { ssr: false }
